refactor QR code scanning in app
